ilm: raid_lock: Don't check the drive state for unlocking
If there have no any valid drive path, e.g. all the drive paths have
been removed from system, the raid locking algorithm will directly
return back without changing the drive state.  For this case, the drive
state will not be transferred to "IDM_INIT" state, so it's not necessary
to use the assertion to check the drive state when unlock the mutex.

Signed-off-by: Leo Yan <leo.yan@linaro.org>